requirements.txt
matplotlib
numpy
pygame
tensorflow
Pillow
quickdraw
tqdm
focal-loss
pydantic
seaborn
scikit-learn
